requirements.txt
# Core dependencies
numpy==1.26.4
Pillow>=11.0.0
pydantic>=2.0

# HuggingFace
datasets==2.19.0
huggingface-hub==0.23.0

# AWS S3
boto3==1.34.110

# Optional
requests==2.31.0
